Added basic toggleable wave output

+++ Source/src/waves.c
@@ -23,7 +23,7 @@ Wave waves[(uint8_t) WAVE_TYPES_COUNT] = {
 };
 
 static void generate_square_waveform(Wave *wave) {
-    for (uint16_t sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
+    for (uint sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
         wave->samples[sample] = (sample < SAMPLES_PER_PERIOD / 2) ? PWM_OUTPUT_RESOLUTION : 0;
     }
 }
@@ -31,7 +31,7 @@ static void generate_square_waveform(Wave *wave) {
 static void generate_triangle_waveform(Wave *wave) {
     // ToDo: Implement
 
-    for (uint16_t sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
+    for (uint sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
         wave->samples[sample] = 0;
     }
 }
@@ -39,7 +39,7 @@ static void generate_triangle_waveform(Wave *wave) {
 static void generate_sawtooth_waveform(Wave *wave) {
     // ToDo: Implement
 
-    for (uint16_t sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
+    for (uint sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
         wave->samples[sample] = 0;
     }
 }
@@ -47,7 +47,7 @@ static void generate_sawtooth_waveform(Wave *wave) {
 static void generate_sine_waveform(Wave *wave) {
     // ToDo: Implement
 
-    for (uint16_t sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
+    for (uint sample = 0; sample < SAMPLES_PER_PERIOD; sample++) {
         wave->samples[sample] = 0;
     }
 }
